% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_alpha.R
\name{alpha_from_data_lineup}
\alias{alpha_from_data_lineup}
\title{Create a plot for visual estimation of alpha}
\usage{
alpha_from_data_lineup(
  c = m/K,
  m = 19,
  K = 30,
  alphas = 10^seq(-3, 2, 0.05),
  ...
)
}
\arguments{
\item{c}{The number of selections a panel must have to be interesting (can be non-integer)}

\item{m}{The number of null panels in the lineup}

\item{K}{The total number of null panel selections (or, in a Rorschach lineup, the total number of evaluations)}

\item{alphas}{Numeric vector of alpha values to conduct simulations for}

\item{...}{additional arguments to sim_interesting_panels}
}
\description{
Create a plot for visual estimation of alpha
}
\examples{
alpha_from_data_lineup()
}
