% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_ml.R
\name{alpha_ml}
\alias{alpha_ml}
\title{Estimate alpha of a symmetric Dirichlet distribution}
\usage{
alpha_ml(p, weight = NULL, eps = 10^(-7))
}
\arguments{
\item{p}{(matrix) of probabilities. Number of columns is interpreted as dimension of the Dirichlet distribution
Number of rows is repetitions.}

\item{weight}{vector}

\item{eps}{error allowed in finding root}
}
\value{
rate parameter alpha of symmetric (flat) Dirichlet distribution
}
\description{
The estimation of the rate parameter alpha is done with Maximum Likelihood Estimation.
}
\examples{
x <- matrix(rgamma(n=2000, shape=.1), ncol=20)
plot(density(x[,1]))
x <- x/rowSums(x)
alpha_ml(x)
shapes <- seq(0.01, 0.75, by = 0.01)
ests <- sapply(shapes, function(s) {
  x <- matrix(rgamma(n=2000, shape=s), ncol=20)
  alpha_ml(x)
})
plot(shapes,ests)
}
