\name{dV}
\alias{dV}
\title{Theoretical distribution for lineups under different scenarios}
\usage{
  dV(x, K, m, scenario, type = "numeric")

  dV(x, K, m, scenario, type = "numeric")
}
\arguments{
  \item{x}{vector of the number of picks of the data plot
  out of K evaluations}

  \item{K}{number of evaluations}

  \item{m}{size of the lineup}

  \item{scenario}{which scenario should be used? 1, 2, or
  3?}

  \item{type}{one of "mpfr" or "numeric". Should the result
  be in arbitrary numeric length or be a numeric?
  Internally the Rmpfr package is used to get maximal
  precision.}

  \item{x}{vector of the number of picks of the data plot
  out of K evaluations}

  \item{K}{number of evaluations}

  \item{m}{size of the lineup}

  \item{scenario}{which scenario should be used? 1, 2, or
  3?}

  \item{type}{one of "mpfr" or "numeric". Should the result
  be in arbitrary numeric length or be a numeric?
  Internally the Rmpfr package is used to get maximal
  precision.}
}
\description{
  Some more details to be written later

  Some more details to be written later
}

