% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.r
\name{dVsim}
\alias{dVsim}
\title{Simulation based density for visual inference}
\usage{
dVsim(x, K, m = 20, N = 10000, scenario = 3, alpha = 1,
  target = 1)
}
\arguments{
\item{x}{number of observed picks of the data plot}

\item{K}{number of evaluations of the same lineup}

\item{m}{size of the lineup}

\item{N}{MC parameter: number of replicates on which MC probabilities are based. Higher number of replicates will decrease MC variability.}

\item{scenario}{numeric value, one of 1, 2, or 3, indicating the type of simulation used: scenario 3 assumes that the same lineup is shown in all K evaluations}

\item{alpha}{positive value, indicates rate parameter used in dirichlet distribution}

\item{target}{location of target plot(s). By default 1. If several targets are present, specify vector of target locations.}
}
\value{
simulation based density to observe x picks of the data plot in K evaluation under the assumption that the data plot is consistent with the null hypothesis. For comparison a p value based on a binomial distribution is provided as well.
}
\description{
Probablity of observing exactly x picks of the data plot in K evaluations of a lineup of size m.
We distinguish between three different scenarios:
\itemize{
\item Scenario I: in each of K evaluations a different data set and a different set of (m-1) null plots is shown.
\item Scenario II: in each of K evaluations the same data set but a different set of (m-1) null plots is shown.
\item Scenario III: the same lineup, i.e. same data and same set of null plots, is shown to K different observers.
}
}
\examples{
dVsim(2, 20, m=3) # triangle test

\dontrun{
## points in red are binomial distribution, black points are for inference
## in lineups using scenario 3
require(ggplot2)
qplot(x=x, y=scenario3, data=dVsim(0:6,6,m=2)) +
   geom_point(aes(x,y=binom), colour="red") + ylim(c(0,0.5))
qplot(x=x, y=scenario3, data=dVsim(0:6,6,m=3)) +
   geom_point(aes(x,y=binom), colour="red") + ylim(c(0,0.5))
}

# lineup with two targets: what are the probabilities to identify at least
# one of the targets?
dVsim(0:5, K=5, m=20, N=10000, scenario=3, target=1:2)
# slight difference between this distribution and the distribution for a
# lineup of size 10 with a single target:
dVsim(0:5, K=5, m=10, N=10000, scenario=3, target=1)
}
