% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_alpha.R
\name{estimate_alpha_visual}
\alias{estimate_alpha_visual}
\title{Create a plot for visual estimation of alpha}
\usage{
estimate_alpha_visual(
  c = m0/K,
  m0 = 19,
  K = 30,
  alphas = 10^seq(-3, 2, 0.05),
  ...
)
}
\arguments{
\item{c}{The number of selections a panel must have to be interesting (can be non-integer)}

\item{m0}{The number of null panels in the lineup}

\item{K}{The total number of null panel selections (or, in a Rorschach lineup, the total number of evaluations)}

\item{alphas}{Numeric vector of alpha values to conduct simulations for}

\item{...}{additional arguments to sim_interesting_panels}
}
\description{
The visual estimation plot contains the expected number of interesting 
panels for each alpha value under the specified experimental conditions and
simulated lineup experiments under each scenario which provide some 
information about the expected variability.
}
\examples{
estimate_alpha_visual()
}
