% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_alpha.R
\name{expected_number_panels}
\alias{expected_number_panels}
\title{Compute the expected number of c-interesting panels for a lineup experiment}
\usage{
expected_number_panels(alpha, c = m0/K, m0 = 19, K = 30)
}
\arguments{
\item{alpha}{The Dirichlet parameter which is related to the number of interesting panels}

\item{c}{The number of selections a panel must have to be interesting (can be non-integer)}

\item{m0}{The number of null panels in the lineup}

\item{K}{The total number of null panel selections (or, in a Rorschach lineup, the total number of evaluations)}
}
\description{
Compute the expected number of c-interesting panels for a lineup experiment
}
\examples{
expected_number_panels(alpha = 1, m0 = 19, K = 30)
}
