% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pvalues.r
\name{h}
\alias{h}
\title{Explicit density function of visual inference under scenario 3 for m = 2}
\usage{
h(x, K)
}
\arguments{
\item{x}{}

\item{K}{}
}
\description{
Explicit density function of visual inference under scenario 3 for m = 2
}
\examples{
h(0:5, K = 5)
## compare to
}

