\name{hdensity}
\alias{hdensity}
\title{Theoretical density for lineups under scenario 4 for m = 2,3}
\usage{
  hdensity(x, K, m)
}
\arguments{
  \item{x}{vector of the number of picks of the data plot
  out of K evaluations}

  \item{K}{number of evaluations}

  \item{m}{size of the lineup}
}
\description{
  Some more details to be written later
}
\examples{
hdensity(0:5, 5, m=2)
## compare to
dvisual(0:5, 5, m=2)
}

