% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.r
\name{lineup}
\alias{lineup}
\title{Simulate a lineup evaluation}
\usage{
lineup(m = 20, K = 1, alpha = 1, dataprob = NULL, nulls = NULL)
}
\arguments{
\item{m}{size of the lineup, typically m=20}

\item{K}{number of independent lineup evaluations}

\item{alpha}{strictly positive numeric value, rate parameter of the corresponding Dirichlet distribution}

\item{dataprob}{positive value describing the interestingness of the data plot. If not specified, a Gamma distributed value is generated}

\item{nulls}{vector of length m (Rorschach) or m-1 (if dataprob is specified) describing the interestingness of the null plots. Simulated from independent Gamma distributions with shape parameter alpha if not specified.}
}
\description{
Simulate a lineup evaluation
}
\examples{
# typical Rorschach plot, evaluated 15 times
x <- lineup(K=15, alpha = 0.15, dataprob=2) 
x
table(x)
alpha_ml(as.numeric(table(x))) # should be close to specified alpha
}
