% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_alpha.R
\name{observed_band}
\alias{observed_band}
\title{Add a band corresponding to the observed number of interesting panels to the visual estimation plot}
\usage{
observed_band(obs, limits, c = m0/K, m0 = 19, K = 30)
}
\arguments{
\item{obs}{Observed number of "c-interesting" panels (panels with at least c selections). If NA, the line is omitted.}

\item{limits}{Lower and upper bounds for number of panels. If NA, the bands are omitted.}

\item{c}{The number of selections a panel must have to be interesting (can be non-integer)}

\item{m0}{The number of null panels in the lineup}

\item{K}{The total number of null panel selections (or, in a Rorschach lineup, the total number of evaluations)}
}
\description{
This function adds a band to assist in the visual estimation of alpha from evaluated lineups.
}
\examples{
estimate_alpha_visual() + observed_band(obs = NA, limits = c(5, 7))
estimate_alpha_visual() + observed_band(obs = 6, limits = NA)
estimate_alpha_visual(c = 1, m0 = 19, K = 50) + observed_band(obs = 6, limits = c(4, 8), c = 1, m0 = 19, K = 50)
}
