% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.r
\name{qV}
\alias{qV}
\title{Theoretical quantile for lineups under different scenarios}
\usage{
qV(q, K, m, scenario, type = "numeric")
}
\arguments{
\item{q}{(vector) of the quantiles for picks of the data plot out of K evaluations}

\item{K}{number of evaluations}

\item{m}{size of the lineup}

\item{scenario}{which scenario should be used? 1, 2, or 3?}

\item{type}{one of "mpfr" or "numeric". Should the result be in arbitrary numeric length or be a numeric? Internally the Rmpfr package is used to get maximal precision.}
}
\description{
Some more details to be written later
}
\examples{
## get critical values of visual triangle test:
qV(q=c(0.95, 0.99), K=c(5,10,15,20, 25, 30), m=3, scenario=1:3)

## get critical values of full lineup test:
qV(q=c(0.95, 0.99), K=c(5,10,15,20, 25, 30), m=20, scenario=3)
}
