% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_alpha.R
\name{rVis}
\alias{rVis}
\title{Functions to produce visual estimate of alpha
Simulate data from a lineup evaluation experiment using the Dirichlet-Multinomial model}
\usage{
rVis(alpha, m = 20, K = 22, N = 50)
}
\arguments{
\item{alpha}{The (scalar) symmetric Dirichlet parameter which is related to the number of interesting panels}

\item{m}{The number of panels in the lineup}

\item{K}{The total number of null panel selections (or, in a Rorschach lineup, the total number of evaluations)}

\item{N}{Number of lineups to simulate}
}
\description{
This function returns panel selection counts simulated from the Dirichlet-
Multinomial model; that is, the result is a \eqn{m \times N} matrix of panel selection counts.
}
\examples{
rVis(alpha = .5, m = 20, K = 30, N = 5)
}
