% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_alpha.R
\name{sim_interesting_panels}
\alias{sim_interesting_panels}
\title{Simulate the number of c-interesting panels for a lineup experiment}
\usage{
sim_interesting_panels(
  alphas = 10^seq(-2, 2, 0.05),
  c = m0/K,
  m0 = 19,
  K = 30,
  N_points = 10,
  avg_n_sims = 10
)
}
\arguments{
\item{alphas}{Numeric vector of alpha values to conduct simulations for}

\item{c}{The number of selections a panel must have to be interesting (can be non-integer)}

\item{m0}{The number of null panels in the lineup}

\item{K}{The total number of null panel selections (or, in a Rorschach lineup, the total number of evaluations)}

\item{N_points}{The number of points to simulate for each value of alpha}

\item{avg_n_sims}{The number of simulations to average to get a single point value.
Averaging several simulations reduces the visual noise but also decreases 
the separation between possible values for a more continuous appearance.}
}
\description{
Simulate the number of c-interesting panels for a lineup experiment
}
\examples{
sim_interesting_panels()
}
