% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pvalues.r
\docType{data}
\name{T1m}
\alias{T1m}
\title{List of coefficients in hdensity}
\usage{
data(T1m)
}
\description{
List of theoretical coefficients to evaluate hdensity in cases m= 2, 3 and 20
}
\details{
List of internally used coefficients to evaluate hdensity
}

