% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi.r
\name{vmulti}
\alias{dmulti}
\alias{pmulti}
\alias{qmulti}
\alias{vmulti}
\title{Simulation based density and distribution functions of visual inference under a multiple choice lineup}
\usage{
dmulti(x, K, k, m = 20, type = "scenario3", N = 5000)

pmulti(x, K, k, m = 20, type = "scenario3", N = 5000)

qmulti(q, K, k, m = 20, type = "scenario3", N = 5000)
}
\arguments{
\item{x}{(vector) of the number of observed data picks}

\item{K}{integer value, specifying the number of participants/evaluations}

\item{k}{(vector) of integer values of the number of observed picks in a multiple choice lineup evaluation}

\item{m}{lineup size, defaults to m=20}

\item{type}{character, one of "scenario1", "scenario2", or "scenario3"}

\item{N}{integer value, number of simulations.}

\item{q}{(vector) of the number of observed data picks}
}
\value{
list consisting of two named items:
\itemize{
\item dmulti: density estimates and vector k
\item pmulti: distribution estimates and vector k
\item qmulti: quantile estimates and vector k
}
}
\description{
Simulation based p value to observe x or more picks of the data plot in K evaluations of a multiple choice lineup test  under the assumption that the data plot is consistent with the null hypothesis.
We distinguish between three different scenarios:
\itemize{
\item Scenario I: in each of K evaluations a different data set and a different set of (m-1) null plots is shown.
\item Scenario II: in each of K evaluations the same data set but a different set of (m-1) null plots is shown.
\item Scenario III: the same lineup, i.e. same data and same set of null plots, is shown to K different observers.
}
}
\examples{
k=rpois(5,lambda=1)+1
m=20
dmulti(0:5,K=5,k=k, type="scenario1", m)
## compare to Poisson Binomial:
require(poibin)
dpoibin(0:5, pp=k/m)

qmulti(c(0.95, 0.99), K=5, k=2)

(k <- rpois(5, lambda=1.5)+1)
reps1 <- ldply(1:10, function(x) dmulti(0:5, 5, k, type="scenario1")$density)
reps2 <- ldply(1:10, function(x) dmulti(0:5, 5, k, type="scenario2")$density)
reps3 <- ldply(1:10, function(x) dmulti(0:5, 5, k, type="scenario3")$density)
reps1$type <- "I"
reps2$type <- "II"
reps3$type <- "III"
reps <- rbind(reps1, reps2, reps3)
library(reshape2)
cols <- brewer.pal(8, "Paired")

mr <- melt(reps, measure.vars=1:6)
mr$variable <- as.numeric(gsub("V", "", mr$variable))
mr$type <- factor(mr$type)
mr$offset <- with(mr, 0.1*c(0,-1,1)[as.numeric(type)])
qplot(variable+offset, value, data=mr, alpha=I(0.75), colour=type) +
  geom_point(aes(x=1:6, y=dpoibin(0:5, pp=k*1/m)), pch=1, size=4, inherit.aes=FALSE)+
  xlab("x")+ylab("P(X=x)")+ggtitle(paste(k, collapse=",")) + theme_bw() + scale_colour_manual(values=cols[-c(1,3,5,6,7)])
}

