% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues-deprecated.r
\name{dV}
\alias{dV}
\title{Theoretical density for lineups under different scenarios}
\usage{
dV(x, K, m, scenario, type = "numeric")
}
\arguments{
\item{x}{vector of the number of picks of the data plot out of K evaluations}

\item{K}{number of evaluations}

\item{m}{size of the lineup}

\item{scenario}{which scenario should be used? 1, 2, or 3?}

\item{type}{one of "mpfr" or "numeric". Should the result be in arbitrary numeric length or be a numeric? Internally the Rmpfr package is used to get maximal precision.}
}
\description{
\lifecycle{deprecated}
}
\examples{
dV(0:5, 5, m=2, scenario=3)
## compare to
dVsim(0:5, 5, m=2, scenario=3)

require(ggplot2)
## probabilities can be computed without numeric loss for K=50:
K <- 50
print(qplot(0:K, dV(0:K, K, m=20, scenario=3)));
print(sum(dV(0:K, K, m=20, scenario=3)));
}
