\name{h}
\alias{h}
\title{Explicit density function of visual inference under scenario 4 for m = 2}
\usage{
  h(x, K)
}
\arguments{
  \item{x}{}

  \item{K}{}
}
\description{
  Explicit density function of visual inference under
  scenario 4 for m = 2
}
\examples{
h(0:5, K = 5)
## compare to
dvisual(0:5, K=5, m=2)
}

